src/target-app/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
prometheus-client==0.21.0
paho-mqtt==2.1.0
